フィルタリングGUI、「地物タイプ指定」と「LOD指定」を統合
